move GetUserInput method from AppUIHelpers into InputReader class, and create InputReaderContainer static class to hold instance of InputReader class (where in unit test we can create a new class that inherits from InputReader and override its methods)
